extract layer sorting to the helper file
